Replace celery with Google PubSub sdk in worker
